Added comment box to AddQRCodeFragment
